change Long type to Int

+++ longpoll/src/commonMain/kotlin/ru/krindra/vkkt/longpoll/bot/updates/objects/AudioNewObject.kt
@@ -7,10 +7,10 @@ import kotlinx.serialization.Serializable
 data class AudioNewObject(
     @SerialName("date") val date: Int? = null,
     @SerialName("artist") val artist: String,
-    @SerialName("album_id") val albumId: Long? = null,
-    @SerialName("owner_id") val ownerId: Long,
+    @SerialName("album_id") val albumId: Int? = null,
+    @SerialName("owner_id") val ownerId: Int,
     @SerialName("duration") val duration: Int,
-    @SerialName("id") val id: Long,
+    @SerialName("id") val id: Int,
     @SerialName("access_key") val accessKey: String? = null,
     @SerialName("title") val title: String,
     @SerialName("performer") val performer: String? = null,
